# contracts/http/analysis-service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UrbanInsight Analysis Service",
    "version": "0.1.0",
    "x-service-name": "analysis-service"
  },
  "paths": {
...
    "/history": {
      "get": {
        "summary": "Get History",
        "operationId": "get_history_history_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryItem"
                  },
                  "title": "Response Get History History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/history/{record_id}": {
      "delete": {
        "summary": "Delete History",
        "operationId": "delete_history_history__record_id__delete",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/history/{record_id}/report": {
      "get": {
        "summary": "Get History Report",
        "operationId": "get_history_report_history__record_id__report_get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          },
          {
            "name": "use_llm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "是否使用LLM(1/0)，未配置key时自动降级",
              "default": 1,
              "title": "Use Llm"
            },
            "description": "是否使用LLM(1/0)，未配置key时自动降级"
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "强制刷新(1/0)，忽略已存报告",
              "default": 0,
              "title": "Refresh"
            },
            "description": "强制刷新(1/0)，忽略已存报告"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "upload_time": {
            "type": "string",
            "format": "date-time",
            "title": "Upload Time"
          },
          "pedestrian_count": {
            "type": "integer",
            "title": "Pedestrian Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "upload_time",
          "pedestrian_count"
        ],
        "title": "HistoryItem"
      },
      "ReportResponse": {
        "properties": {
          "record_id": {
            "type": "integer",
            "title": "Record Id"
          },
          "llm_used": {
            "type": "boolean",
            "title": "Llm Used"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          }
        },
        "type": "object",
        "required": [
          "record_id",
          "llm_used",
          "report"
        ],
        "title": "ReportResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}